+++ backend/workouts/__tests__/integration/api/workout.spec.ts
@@ -5,28 +5,14 @@ import {User, Workout} from '@entities';
 import MockDate from 'mockdate';
 import {decode} from 'jsonwebtoken';
 import {AuthenticationError, WorkoutNotFoundError} from '@errors';
-import {createToken} from './utils';
+import {createAdminUserAndToken, createUserAndToken} from './utils';
 
 /**
- * Workout APIs
+ * Test Groups
  *
  * @group integration
  * @group workout
  */
-
-async function createUserAndToken(): Promise<[User, string]> {
-    const userId = randomUUID();
-    const accessToken = await createToken({userId, roles: ['user']});
-    const user = await User.create({id: userId}).save();
-    return [user, accessToken];
-}
-
-async function createAdminUserAndToken(): Promise<[User, string]> {
-    const userId = randomUUID();
-    const accessToken = await createToken({userId, roles: ['admin']});
-    const user = await User.create({id: userId}).save();
-    return [user, accessToken];
-}
 
 describe('admin workout apis', () => {
     const url = '/api/v1/workouts/';
